boolean->bool, double->float, short->int, fix class constant defs

+++ src/org/ggp/base/player/event/PlayerTimeEvent.java
@@ -5,13 +5,13 @@ import org.ggp.base.util.observer.Event;
 class PlayerTimeEvent(Event):
 {
 
-    time = long()
+    time = int()
 
-    def PlayerTimeEvent(time=long())
+    def PlayerTimeEvent(time=int())
 	{
         this.time = time;
 
-    def long getTime()
+    def int getTime()
 	{
         return time;
