+++ src/org/ggp/base/player/request/grammar/PlayRequest.java
@@ -25,7 +25,7 @@ class PlayRequest(Request):
     def getMatchId():  # String
         return matchId;
 
-    def String process(long receptionTime)
+    def String process(int receptionTime)
 	{
 	    // First, check to ensure that this play request is for the match
 	    // we're currently playing. If we're not playing a match, or we're
