+++ src/org/ggp/base/player/request/grammar/PreviewRequest.java
@@ -20,7 +20,7 @@ class PreviewRequest(Request):
     def getMatchId():  # String
         return null;
 
-    def String process(long receptionTime)
+    def String process(int receptionTime)
 	{
 		// Ensure that we aren't already playing a match. If we are,
 	    // ignore the message, saying that we're busy.
